Extract contract into separate module file
